# Adjust chain.runtime_events indexes