@pro-wh 2023-08-15
oh it's because the query wants some quite sophisticated ordering. the index is only sorted down to round, so it doesn't know about how the tx_index and event type are ordered within that. so it's sorting all matching events, ever