@mitjat 2023-08-15
> oh it's because the query wants some quite sophisticated ordering. the index is only sorted down to round, so it doesn't know about how the tx_index and event type are ordered within that. so it's sorting all matching events, ever

But "all matching events" in this case is ... only 348 of them (`rows=348`), no? And it takes about 150ms (`actual time=63.521..216.633`) to sort them, somehow. I _must_ be reading something wrong :)