@pro-wh 2023-08-16
no it's this one from the events side of the join

```
               ->  Incremental Sort  (cost=58.92..1358075.02 rows=115438 width=663) (actual time=35.149..95.806 rows=116 loops=3)
                     Sort Key: evs.round DESC, evs.tx_index, evs.type, ((evs.body)::text)
                     Presorted Key: evs.round
```