@Andrew7234 2023-08-17
Thanks for the thorough testing! 

>  on prod mainnet, both my query and the original query take forever

Hm the index wasn't applied to either production or staging yet, so it makes sense that the original query is slow. Seems like your optimized query would also benefit from the index, so I tested it out. 

Original (current) query with new `(runtime, evm_log_signature, round)` index:
```
                                                                                                      QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1048.62..6507.85 rows=1000 width=698) (actual time=27.270..125.575 rows=349 loops=1)
   ->  Nested Loop Left Join  (cost=1048.62..1720913.73 rows=315038 width=698) (actual time=27.268..125.473 rows=349 loops=1)
         Join Filter: ((evs.runtime = tokens.runtime) AND ((preimages.address)::text = (tokens.token_address)::text))
         Rows Removed by Join Filter: 3794
         ->  Gather Merge  (cost=1044.11..1557825.06 rows=315038 width=675) (actual time=27.169..121.557 rows=349 loops=1)
               Workers Planned: 2
               Workers Launched: 2
               ->  Incremental Sort  (cost=44.09..1520461.83 rows=131266 width=675) (actual time=0.693..17.134 rows=116 loops=3)
                     Sort Key: evs.round DESC, evs.tx_index, evs.type, ((evs.body)::text)
                     Presorted Key: evs.round
                     Full-sort Groups: 2  Sort Method: quicksort  Average Memory: 89kB  Peak Memory: 89kB
                     Worker 0:  Full-sort Groups: 2  Sort Method: quicksort  Average Memory: 89kB  Peak Memory: 89kB
                     Worker 1:  Full-sort Groups: 4  Sort Method: quicksort  Average Memory: 110kB  Peak Memory: 132kB
                       Pre-sorted Groups: 27  Sort Method: quicksort  Average Memory: 36kB  Peak Memory: 277kB
                     ->  Nested Loop Left Join  (cost=1.00..1515508.20 rows=131266 width=675) (actual time=0.073..1.306 rows=116 loops=3)
                           ->  Parallel Index Scan Backward using ix_runtime_events_evm_log_signature on runtime_events evs  (cost=0.56..569171.72 rows=131266 width=628) (actual time=0.027..0.142 rows=116 loops=3)
                                 Index Cond: ((runtime = 'sapphire'::runtime) AND (evm_log_signature = '\xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef'::bytea))
                           ->  Index Scan using ix_address_preimages_address_data on address_preimages preimages  (cost=0.43..7.20 rows=1 width=68) (actual time=0.004..0.004 rows=1 loops=349)
                                 Index Cond: (address_data = decode((evs.body ->> 'address'::text), 'base64'::text))
                                 Filter: ((context_identifier = 'oasis-runtime-sdk/address: secp256k1eth'::text) AND ((context_version)::integer = 0))
         ->  Materialize  (cost=4.50..56.57 rows=29 width=65) (actual time=0.000..0.002 rows=12 loops=349)
               ->  Bitmap Heap Scan on evm_tokens tokens  (cost=4.50..56.43 rows=29 width=65) (actual time=0.041..0.062 rows=29 loops=1)
                     Recheck Cond: (runtime = 'sapphire'::runtime)
                     Heap Blocks: exact=5
                     ->  Bitmap Index Scan on evm_tokens_pkey  (cost=0.00..4.50 rows=29 width=0) (actual time=0.029..0.029 rows=29 loops=1)
                           Index Cond: (runtime = 'sapphire'::runtime)
 Planning Time: 0.609 ms
 Execution Time: 125.859 ms
(28 rows)
```

Optimized query with index: 
```
                                                                                              QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Hash Left Join  (cost=71.89..10167.61 rows=1000 width=666) (actual time=0.320..21.870 rows=349 loops=1)
   Hash Cond: ((preimages.address)::text = (tokens.token_address)::text)
   ->  Nested Loop Left Join  (cost=17.50..10105.60 rows=1000 width=671) (actual time=0.261..21.508 rows=349 loops=1)
         ->  Limit  (cost=17.07..1878.60 rows=1000 width=656) (actual time=0.242..19.379 rows=349 loops=1)
               ->  Incremental Sort  (cost=17.07..586471.23 rows=315038 width=656) (actual time=0.241..19.320 rows=349 loops=1)
                     Sort Key: evs.round DESC, evs.tx_index, evs.type, ((evs.body)::text)
                     Presorted Key: evs.round
                     Full-sort Groups: 8  Sort Method: quicksort  Average Memory: 62kB  Peak Memory: 67kB
                     Pre-sorted Groups: 9  Sort Method: quicksort  Average Memory: 38kB  Peak Memory: 140kB
                     ->  Index Scan Backward using ix_runtime_events_evm_log_signature on runtime_events evs  (cost=0.56..572584.63 rows=315038 width=656) (actual time=0.030..3.430 rows=349 loops=1)
                           Index Cond: ((runtime = 'sapphire'::runtime) AND (evm_log_signature = '\xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef'::bytea))
         ->  Index Scan using ix_address_preimages_address_data on address_preimages preimages  (cost=0.43..8.21 rows=1 width=68) (actual time=0.004..0.004 rows=1 loops=349)
               Index Cond: (address_data = decode((evs.body ->> 'address'::text), 'base64'::text))
               Filter: ((context_identifier = 'oasis-runtime-sdk/address: secp256k1eth'::text) AND ((context_version)::integer = 0))
   ->  Hash  (cost=54.02..54.02 rows=29 width=61) (actual time=0.051..0.053 rows=29 loops=1)
         Buckets: 1024  Batches: 1  Memory Usage: 11kB
         ->  Index Scan using evm_tokens_pkey on evm_tokens tokens  (cost=0.28..54.02 rows=29 width=61) (actual time=0.010..0.036 rows=29 loops=1)
               Index Cond: (runtime = 'sapphire'::runtime)
 Planning Time: 0.643 ms
 Execution Time: 21.993 ms
(20 rows)
```

The optimized query(20-40ms) is consistently faster than the original (90-120ms). I'm on the fence as to whether to include the updated query in this PR - the optimized query is more complex but the performance speedup is certainly noticable so I think overall I'm leaning towards it.

> same query went from 3 minutes to 1 second after rerunning a few times 

Probably the postgres cache. I confirmed that both queries slow significantly (> 1min) after removing the index.